fix: change click event from refresh to refresh()
The following is observed

Type '(opts?: AsyncDataExecuteOptions | undefined) => Promise<void>' is not assignable to type '(payload: MouseEvent) => void'

Reason being:
When using @click="refresh" It expects refresh to be a function that takes a MouseEvent as its argument.

Using the following @click="refresh()"
calls refresh immediately and assigns the return value to the click handler. This effectively bypasses the type mismatch